[USER] Backend ROLE + USER => REPOSITORY

+++ tms/src/main/java/com/netcracker/edu/tms/model/Role.java
@@ -6,6 +6,7 @@ import lombok.NoArgsConstructor;
 import org.hibernate.annotations.NaturalId;
 
 import javax.persistence.*;
+import java.math.BigInteger;
 
 @Data
 @AllArgsConstructor
@@ -17,7 +18,7 @@ public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", unique = true, nullable = false)
-    private int id;
+    private BigInteger id;
 
     @Column(name="name", unique = true, nullable = false)
     private String name;
